making room components
