+++ package.json
@@ -12,6 +12,9 @@
         "@types/react": "^18.3.0",
         "@types/react-dom": "^18.3.0",
         "axios": "^1.6.8",
+        "dotenv": "^16.4.5",
+        "nth-check": "^2.1.1",
+        "postcss": "^8.4.38",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-router-dom": "^6.23.0",
