+++ src/index.css
@@ -1,7 +1,7 @@
 :root {
   --main-bg-color1: #00948B;
   --main-bg-color2: #004A45;
-  --main-text-color: white;
+  /* --main-text-color: white; */
 }
 
 body {
@@ -15,5 +15,5 @@ body {
   /* background: linear-gradient(to bottom, var(--main-bg-color2), var(--main-bg-color1) 30%, var(--main-bg-color2)) no-repeat; */
   background: var(--main-bg-color2) no-repeat;
   color: var(--main-text-color);
-  text-align: center;
+  /* text-align: center; */
 }
