+++ src/server/server.js
@@ -1,5 +1,6 @@
 import {io} from "socket.io-client"
 
-const socket = io("http://localhost:5001")
+
+const socket = io("http://192.168.0.146:5001")
 
 export default socket;
